refactor: TanStack Query 로직 분리
